+++ test/package_meta_test.dart
@@ -48,19 +48,6 @@ void main() {
       expect(p.version, isNotNull);
     });
 
-    test('has a description', () {
-      expect(
-          // ignore: deprecated_member_use_from_same_package
-          p.description,
-          equals(
-              'A non-interactive HTML documentation generator for Dart source code.'));
-    });
-
-    test('has a repository', () {
-      // ignore: deprecated_member_use_from_same_package
-      expect(p.repository, equals('https://github.com/dart-lang/dartdoc'));
-    });
-
     test('is valid', () {
       expect(p.isValid, isTrue);
       expect(p.getInvalidReasons(), isEmpty);
@@ -73,26 +60,6 @@ void main() {
               .readAsMalformedAllowedStringSync(p.getReadmeContents()!),
           contains(
               'Use `dart doc` to generate HTML documentation for your Dart package.'));
-    });
-
-    test('has a license', () {
-      // ignore: deprecated_member_use_from_same_package
-      expect(p.getLicenseContents(), isNotNull);
-      expect(
-          resourceProvider
-              // ignore: deprecated_member_use_from_same_package
-              .readAsMalformedAllowedStringSync(p.getLicenseContents()!),
-          contains('Copyright 2014, the Dart project authors.'));
-    });
-
-    test('has a changelog', () {
-      // ignore: deprecated_member_use_from_same_package
-      expect(p.getChangelogContents(), isNotNull);
-      expect(
-          resourceProvider
-              // ignore: deprecated_member_use_from_same_package
-              .readAsMalformedAllowedStringSync(p.getChangelogContents()!),
-          contains('## 0.2.2'));
     });
   });
 
@@ -114,14 +81,6 @@ void main() {
       expect(p.version, isNotNull);
     });
 
-    test('has a description', () {
-      expect(
-          // ignore: deprecated_member_use_from_same_package
-          p.description,
-          equals(
-              'The Dart SDK is a set of tools and libraries for the Dart programming language.'));
-    });
-
     test('has a homepage', () {
       expect(p.homepage, equals('https://github.com/dart-lang/sdk'));
     });
@@ -132,11 +91,6 @@ void main() {
           resourceProvider
               .readAsMalformedAllowedStringSync(p.getReadmeContents()!),
           contains('Welcome to the'));
-    });
-
-    test('does not have a license', () {
-      // ignore: deprecated_member_use_from_same_package
-      expect(p.getLicenseContents(), isNull);
     });
   });
 }